+++ examples/scripts/BPX_spm.py
@@ -46,12 +46,12 @@ cost = pybop.SumSquaredError(problem)
 optim = pybop.CMAES(cost, max_iterations=40)
 
 # Run the optimisation
-x, final_cost = optim.run()
+results = optim.run()
 print("True parameters:", parameters.true_value())
-print("Estimated parameters:", x)
+print("Estimated parameters:", results.x)
 
 # Plot the timeseries output
-pybop.quick_plot(problem, problem_inputs=x, title="Optimised Comparison")
+pybop.quick_plot(problem, problem_inputs=results.x, title="Optimised Comparison")
 
 # Plot convergence
 pybop.plot_convergence(optim)